DTFS2-4516 - move deal-submit-emails api test, add GEF AIN/MIN email call tests
